+++ server/runtime_lua_nakama.go
@@ -5992,7 +5992,9 @@ func (n *RuntimeLuaNakamaModule) tournamentCreate(l *lua.LState) int {
 		return 0
 	}
 
-	sortOrder := l.OptString(2, "desc")
+	authoritative := l.OptBool(2, true)
+
+	sortOrder := l.OptString(3, "desc")
 	var sortOrderNumber int
 	switch sortOrder {
 	case "asc":
@@ -6000,11 +6002,11 @@ func (n *RuntimeLuaNakamaModule) tournamentCreate(l *lua.LState) int {
 	case "desc":
 		sortOrderNumber = LeaderboardSortOrderDescending
 	default:
-		l.ArgError(2, "expects sort order to be 'asc' or 'desc'")
+		l.ArgError(3, "expects sort order to be 'asc' or 'desc'")
 		return 0
 	}
 
-	operator := l.OptString(3, "best")
+	operator := l.OptString(4, "best")
 	var operatorNumber int
 	switch operator {
 	case "best":
@@ -6016,25 +6018,25 @@ func (n *RuntimeLuaNakamaModule) tournamentCreate(l *lua.LState) int {
 	case "decr":
 		operatorNumber = LeaderboardOperatorDecrement
 	default:
-		l.ArgError(3, "expects sort order to be 'best', 'set', 'decr' or 'incr'")
+		l.ArgError(4, "expects sort order to be 'best', 'set', 'decr' or 'incr'")
 		return 0
 	}
 
-	duration := l.OptInt(4, 0)
+	duration := l.OptInt(5, 0)
 	if duration <= 0 {
-		l.ArgError(4, "duration must be > 0")
+		l.ArgError(5, "duration must be > 0")
 		return 0
 	}
 
-	resetSchedule := l.OptString(5, "")
+	resetSchedule := l.OptString(6, "")
 	if resetSchedule != "" {
 		if _, err := cronexpr.Parse(resetSchedule); err != nil {
-			l.ArgError(5, "expects reset schedule to be a valid CRON expression")
+			l.ArgError(6, "expects reset schedule to be a valid CRON expression")
 			return 0
 		}
 	}
 
-	metadata := l.OptTable(6, nil)
+	metadata := l.OptTable(7, nil)
 	metadataStr := "{}"
 	if metadata != nil {
 		metadataMap := RuntimeLuaConvertLuaTable(metadata)
@@ -6046,36 +6048,36 @@ func (n *RuntimeLuaNakamaModule) tournamentCreate(l *lua.LState) int {
 		metadataStr = string(metadataBytes)
 	}
 
-	title := l.OptString(7, "")
-	description := l.OptString(8, "")
-	category := l.OptInt(9, 0)
+	title := l.OptString(8, "")
+	description := l.OptString(9, "")
+	category := l.OptInt(10, 0)
 	if category < 0 || category >= 128 {
-		l.ArgError(9, "category must be 0-127")
+		l.ArgError(10, "category must be 0-127")
 		return 0
 	}
-	startTime := l.OptInt(10, 0)
+	startTime := l.OptInt(11, 0)
 	if startTime < 0 {
-		l.ArgError(10, "startTime must be >= 0.")
+		l.ArgError(11, "startTime must be >= 0.")
 		return 0
 	}
-	endTime := l.OptInt(11, 0)
+	endTime := l.OptInt(12, 0)
 	if endTime != 0 && endTime <= startTime {
-		l.ArgError(11, "endTime must be > startTime. Use 0 to indicate a tournament that never ends.")
+		l.ArgError(12, "endTime must be > startTime. Use 0 to indicate a tournament that never ends.")
 		return 0
 	}
-	maxSize := l.OptInt(12, 0)
+	maxSize := l.OptInt(13, 0)
 	if maxSize < 0 {
-		l.ArgError(12, "maxSize must be >= 0")
+		l.ArgError(13, "maxSize must be >= 0")
 		return 0
 	}
-	maxNumScore := l.OptInt(13, 0)
+	maxNumScore := l.OptInt(14, 0)
 	if maxNumScore < 0 {
-		l.ArgError(13, "maxNumScore must be >= 0")
+		l.ArgError(14, "maxNumScore must be >= 0")
 		return 0
 	}
-	joinRequired := l.OptBool(14, false)
+	joinRequired := l.OptBool(15, false)
 
-	if err := TournamentCreate(l.Context(), n.logger, n.leaderboardCache, n.leaderboardScheduler, id, sortOrderNumber, operatorNumber, resetSchedule, metadataStr, title, description, category, startTime, endTime, duration, maxSize, maxNumScore, joinRequired); err != nil {
+	if err := TournamentCreate(l.Context(), n.logger, n.leaderboardCache, n.leaderboardScheduler, id, authoritative, sortOrderNumber, operatorNumber, resetSchedule, metadataStr, title, description, category, startTime, endTime, duration, maxSize, maxNumScore, joinRequired); err != nil {
 		l.RaiseError("error creating tournament: %v", err.Error())
 	}
 	return 0